Remove base patch (#71)
Merge after #69 

This deletes the QEDbase patch which is no longer necessary as the
relevant PR in QEDbase was merged.
Unfortunately, there seems to be a somewhat strange interaction of
doubly defined typenames and `@reexport`, creating an issue in QEDbase's
documentation building. This means that I had to remove some `@extref`
statements for now that can be readded later, when QEDbase is ready.